NUEVO MODELO DE DATOS Y AJUSTES DE LOGS Y TARJETAS DE ACCESO

+++ controllers/AdministratorsController.php
@@ -63,31 +63,38 @@ class AdministratorsController extends Controller {
      */
     public function actionCreate() {
         $model = new Administrators();
-        $model->scenario = 'create';
 
         if ($model->load(Yii::$app->request->post())) {
 
             //INSTACIO EL ARCHIVO CARGADO
-            if (!$model->file = \yii\web\UploadedFile::getInstance($model, 'file')) {
-                Yii::$app->session->setFlash('error', "El archivo no pudo "
-                        . "ser cargado. Inténtelo de nuevo.");
-                return $this->redirect(['index']);
+            $model->file = \yii\web\UploadedFile::getInstance($model, 'file');
+            if (!is_null($model->file)) {
+                if (!$model->file = \yii\web\UploadedFile::getInstance($model, 'file')) {
+                    Yii::$app->session->setFlash('error', "El archivo no pudo "
+                            . "ser cargado. Inténtelo de nuevo.");
+                    return $this->redirect(['index']);
+                }
+
+                //RUTA DE ALMACENAJE LOCAL Y NOMBRE
+                $ruta = 'archivos/' . date('YmdHis') . '-'
+                        . strtolower(trim(str_replace($this->especial, $this->wespecial, $model->file->baseName))) . '.'
+                        . strtolower(trim($model->file->extension));
+                if (!@$model->file->saveAs($ruta, false)) {
+                    Yii::$app->session->setFlash('error', "El archivo no pudo "
+                            . "ser guardado. Inténtelo de nuevo.");
+                    return $this->redirect(['index']);
+                }
+
+                //GUARDO LOS DATOS
+                $model->photo = $ruta;
+            } else {
+                $model->photo = "";
             }
 
-            //RUTA DE ALMACENAJE LOCAL Y NOMBRE
-            $ruta = 'archivos/' . date('YmdHis') . '-'
-                    . strtolower(trim(str_replace($this->especial, $this->wespecial, $model->file->baseName))) . '.'
-                    . strtolower(trim($model->file->extension));
-            if (!@$model->file->saveAs($ruta, false)) {
-                Yii::$app->session->setFlash('error', "El archivo no pudo "
-                        . "ser guardado. Inténtelo de nuevo.");
-                return $this->redirect(['index']);
-            }
-
-            //GUARDO LOS DATOS
-            $model->photo = $ruta;
             if (!$model->save()) {
-                unlink($model->photo);
+                if (file_exists($model->photo)) {
+                    unlink($model->photo);
+                }
                 Yii::$app->session->setFlash('error', "El archivo no pudo "
                         . "ser cargado. Inténtelo de nuevo.");
                 return $this->redirect(['index']);
@@ -110,7 +117,7 @@ class AdministratorsController extends Controller {
     public function actionUpdate($id) {
         $model = $this->findModel($id);
 
-         if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
             //INSTACIO EL ARCHIVO CARGADO
             $model->file = \yii\web\UploadedFile::getInstance($model, 'file');
             if (!is_null($model->file)) {
@@ -124,11 +131,13 @@ class AdministratorsController extends Controller {
                     return $this->redirect(['index']);
                 }
                 //GUARDO LOS DATOS
-                unlink($model->photo);
+                if (file_exists($model->photo)) {
+                    unlink($model->photo);
+                }
                 $model->photo = $ruta;
             }
 
-            if (!$model->save()) {                
+            if (!$model->save()) {
                 Yii::$app->session->setFlash('error', "El archivo no pudo "
                         . "ser cargado. Inténtelo de nuevo.");
                 return $this->redirect(['index']);
@@ -150,7 +159,9 @@ class AdministratorsController extends Controller {
      */
     public function actionDelete($id) {
         $photo = $this->findModel($id);
-        unlink($photo->photo);
+        if (file_exists($photo->photo)) {
+            unlink($photo->photo);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
